fullstack-orthanc-dicom/pacs-service/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.12.1

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# HTTP
httpx==0.25.2
requests==2.31.0

# DICOM
pydicom==2.4.4
pynetdicom==2.0.2

# Integration
hl7apy==1.3.4
fhir.resources==7.1.0

# Cloud
boto3==1.34.0
botocore==1.34.0

# Utils
python-dateutil==2.8.2
python-json-logger==2.0.7
python-dotenv==1.0.0
python-slugify==8.0.1
email-validator==2.1.0

# Monitoring & Metrics
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
psutil==5.9.6

# Worker
celery==5.3.4
redis==4.5.1
flower==2.0.1

# Exports
reportlab==4.0.7
weasyprint==60.1
Pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
openpyxl==3.1.2

# AI
google-generativeai==0.3.2
openai>=1.0.0

# Cache
fastapi-cache2[redis]==0.2.2
